requirements.txt
boto3==1.26.60
botocore==1.29.60
certifi==2022.12.7
charset-normalizer==3.0.1
htmlmin==0.1.12
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
pika==1.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
